Bugfix: couldn't set up app due to missing database.yaml
The database.yml file was missing because it had been added to the .gitignore.  I removed it from the .gitignore and created a new yaml file. Now the rake db:create file works.

Analogy: not having the database.yml file is like giving someone a house without giving them the keys.  They have to take the trouble to change the locks before they get inside.