When I press the keycodes listed on a keyboard page, a noise is made
The deveopment of the script can be seen in the commit history of this fiddle: http://jsfiddle.net/tararoys/2Sexn/35/

The difference between this commit and the fiddle is that I needed to call keyboard.init in a document.ready. The keyboard knows which of your keyboard keys to wire up based on the keycodes listed in the key divs. The javascript was trying to read the divs before the page had loaded.