Added hard-coded volume of .5 to every key.
For some reason, the panner, delay, and volume settings are being saved as zeros to the database, so they don't render on the page.  As a result, I can't extract them from the html divs. I therefore hardcoded a volume of .5 on line 39 of keyboard.js instead of using jquery to extract it from the page