Does activerecord let you create an attribute you did not define in the database with the new function?
Nope!

Activerecord attributes are like a toolbox where the tools all are nested in their own space.  For example, if your toolkit has a hammer, there will be a hammer-shaped space that the hammer can fit into exactly. http://amigakit.leamancomputing.com/catalog/images/computer-toolkit-26-piece.jpg   Tools can be missing, but you can't put a completely differet tool in that sort of kit, because there isn't a space for it.