Changed the keys in the database to match our specification
Several of the values in our keys table were listed as integers when
they needed to be listed as floating point numbers.

Here is the specification:

setting    |  data type  | range             |
----------------------------------------------
frequency  |  float      | 50.000 - 5000.000 |
wave_type  |  integer    | 1 - 4             |
panner     |  float      | -3 to 3           |
delay      |  float      | 0 to 1            |
filter     |  float      | -4800 to 4800     |
volume     |  float      | 0 to 1            |
keycode    |  integer    | 0 to 256          |

The database now matches that specification.