Test that frequency must be between 50.00 and 5000.00
Have a test and a validation that forces the input to be between this range. I think it also has the side-benefit of not accepting things that are not numbers.  Yay for sanitizing database input!

This is approximately the range of frequencies on a piano.  The three lowest keys are lower than 50.  The highest key on a piano is lower than 5000.00.