Bugfix got rid of default sound on the 'A' key.
I left testing code bound to the 'A' key, so if you didn't reset the
'A' key, the A key played middle 'C.' It did this on every page,
so pressing the 'A' resulted in a beep being played.  This
was distracting and wierd when you are trying to log in and the
computer beeps at you every time you press the 'a' key.